Replace invalid use of `gettextize`
`gettextize` isn't intended to be used in an automated fashion¹. Switch
to `autopoint`. But, `autopoint` requires a specified version of
`gettext` in `configure.ac` (so add whatever's currently on my system --
not a good general solution).

¹: Last paragraph of
http://www.gnu.org/software/gettext/manual/html_node/gettextize-Invocation.html

"""
It is also important to understand that gettextize is not part of the
GNU build system, in the sense that it should not be invoked
automatically, and not be invoked by someone who doesn’t assume the
responsibilities of a package maintainer. For the latter purpose, a
separate tool is provided, see autopoint Invocation.
"""