do not blink warning message, closes #24
The WARNING message about lack of support in the database had extra
flags on top of "red": it would blink on certain terminals (like
rxvt-unicode). This would make the message really hard to read, which
is probably not the intention.

This simply removes the blinking setting from the terminal escape codes.